backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
requests==2.32.3
python-dotenv==1.0.1
cryptography==44.0.0
openai==1.58.1
beautifulsoup4==4.12.3
lxml==5.3.0
